+++ database/migrations/2022_11_29_071555_create_personal_info_table.php
@@ -13,26 +13,28 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('personal_info', function (Blueprint $table) {
-            $table->id();
-            $table->unsignedBigInteger('user_id');
-            $table->string('first_name');
-            $table->string('middle_name');
-            $table->string('last_name');
-            $table->string('street_address_one');
-            $table->string('street_address_two');
-            $table->string('city');
-            $table->string('state');
-            $table->integer('zip_code');
-            $table->string('cell_phone');
-            $table->string('email');
-            $table->longText('social_links')->nullable();
-            $table->string('parent_email_one');
-            $table->string('parent_email_two');
-            $table->tinyInteger('is_draft')->default(0)->comment('0 => draft, 1 => published');
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('personal_info')) {
+            Schema::create('personal_info', function (Blueprint $table) {
+                $table->id();
+                $table->unsignedBigInteger('user_id');
+                $table->string('first_name');
+                $table->string('middle_name');
+                $table->string('last_name');
+                $table->string('street_address_one');
+                $table->string('street_address_two');
+                $table->string('city');
+                $table->string('state');
+                $table->integer('zip_code');
+                $table->string('cell_phone');
+                $table->string('email');
+                $table->longText('social_links')->nullable();
+                $table->string('parent_email_one');
+                $table->string('parent_email_two');
+                $table->tinyInteger('is_draft')->default(0)->comment('0 => draft, 1 => published');
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
